Feature79 : 게시글 수정버튼, 사진 업로드 삭제 버튼 추가 / 게시글, 댓글 삭제 시 확인창 추가

+++ src/components/community/PostDetail.js
@@ -10,7 +10,7 @@ import LikeIcon from '@material-ui/icons/Favorite';
 import UpdateIcon from '@material-ui/icons/Edit';
 import ReplyCreate from './ReplyCreate';
 import ReplyList from './ReplyList';
-import { LineAxisSharp } from '@mui/icons-material';
+
 
 
 
@@ -73,16 +73,21 @@ const PostDetail = () => {
     }
   };
 
-  const handleDelete = async () => {
-    try {
-      await axios.delete(`/post/delete/${postId}`);
-      alert('게시글이 삭제되었습니다.');
-      navigate('/community/post/list'); // 삭제 후 목록으로 돌아가기
-    } catch (error) {
-      console.error('게시글 삭제 실패:', error);
-      alert('게시글 삭제에 실패했습니다.');
+  const handleDeletePost = async () => {
+    // 사용자로부터 삭제 확인을 받기 위한 알림창 표시
+    const shouldDelete = window.confirm('게시글을 삭제하시겠습니까?');
+    if (shouldDelete) {
+        try {
+            await axios.delete(`/post/delete/${postId}`);
+            alert('게시글이 삭제되었습니다.');
+            navigate('/community/post/list'); // 삭제 후 목록으로 돌아가기
+        } catch (error) {
+            console.error('게시글 삭제 실패:', error);
+            alert('게시글 삭제에 실패했습니다.');
+        }
     }
-  };
+};
+
 
   const handleDeleteReply = (replyId) => {
     // 댓글 삭제 로직을 구현하고, 삭제 후 업데이트된 댓글 목록을 설정
@@ -90,6 +95,17 @@ const PostDetail = () => {
     const updatedReplies = replies.filter(reply => reply.replyId !== replyId);
     setReplies(updatedReplies);
   };  
+
+  const handleUpdateReply = async (replyId, updatedReply) => {
+    try {
+        await axios.put(`/reply/${replyId}`, { content: updatedReply });
+        const responseReplies = await axios.get(`/reply/${postId}/all`);
+        setReplies(responseReplies.data);
+    } catch (error) {
+        console.error('댓글 수정 실패:', error);
+        alert('댓글 수정에 실패했습니다.');
+    }
+  };
 
 
   return (
@@ -191,11 +207,14 @@ const PostDetail = () => {
                     variant="contained"
                     color="primary"
                     component={Link}
-                    onClick={handleDelete}
+                    onClick={handleDeletePost}
                     style={{ marginTop: '10px', marginLeft: '10px' }}
               > 삭제하기 </Button> <br /> <br />
 
-              <ReplyList replies={replies} onDeleteReply={handleDeleteReply}/>
+              <ReplyList 
+                replies={replies} 
+                onDeleteReply={handleDeleteReply}
+                onUpdateReply={handleUpdateReply} />
               <ReplyCreate postId={postId} onReplySubmit={handleNewReply} />
